feat: add pixel art cursor
Add two png files for the cursor, one to
represent the default cursor and the other
one to be a pointer cursor.

Add the default one for the html tag. And
a `inherited` property for label tag.

The pointer one is being added on the button,
radio, checkbox and a tags.