ci: Remove caching (temporarily)
I finally figured it out. Restoring the cache was trying to restore the cache from an older build
that used a different image in which the root user was the one doing all the tasks. I'm pushing a
build that doesn't restore the cache, then I'll add those steps back in after the next build.