build(commitlint): Remove `semantic-commitlint`
This removes `semantic-commitlint` from both the local build pipeline and the CI build pipeline. The
original value of `semantic-commitlint` was that it would handle linting only the relevant commits,
i.e. the commit range under review for a PR. However, the library doesn't seem to respect our
altered `commitlint` rules, so we'll be switching to an alternate strategy to handle linting of
specific ranges.

#13